Optimize tilelang norm_fn with bf16 MFMA
Switch T.gemm from f32 to bf16 operands, triggering
v_mfma_f32_16x16x16_bf16 instead of v_mfma_f32_16x16x4_f32
(4x instruction throughput). Enable num_stages=2 pipeline.

Split fwd_mul into two kernels (GEMM + sqrsum) to avoid
tilelang LayoutInference conflicts between MFMA and
element-wise fragment layouts.

Performance at N=4096, H=16384:
  Before (f32 MFMA): estimated ~1.2ms (f32 path)
  After  (bf16 MFMA): 0.310ms
  Triton bf16 MFMA:   0.328ms
  torch.mm (rocBLAS):  0.660ms

+++ tests/mhc/test_norm_fn_mi300.py
@@ -189,7 +189,7 @@ if __name__ == '__main__':
             fwd_mul_k = _mhc_pre_norm_fn_fwd_mul(24, n_rms, rgs)
             N_t = 64
             x_t = torch.randn(N_t, n_rms * rgs, dtype=torch.bfloat16, device='cuda')
-            fn_t2 = torch.randn(24, n_rms * rgs, dtype=torch.float32, device='cuda')
+            fn_t2 = torch.randn(24, n_rms * rgs, dtype=torch.bfloat16, device='cuda')
             out_t = torch.empty(N_t, n_rms, 24, dtype=torch.float32, device='cuda')
             sq_t = torch.empty(N_t, n_rms, dtype=torch.float32, device='cuda')
             fwd_mul_k(x_t, fn_t2, out_t, sq_t)
